test: 로또 번호를 별도의 테스트로 분리
